[feat] microscopy - edit annotation label
